MuseTalk/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118 
torch==2.0.1
torchvision==0.15.2 
torchaudio==2.0.2
diffusers==0.27.2
accelerate==0.28.0
tensorflow==2.12.0
tensorboard==2.12.0
opencv-python==4.9.0.80
soundfile==0.12.1
transformers==4.39.2
huggingface_hub==0.20.2

gdown
requests
imageio[ffmpeg]

omegaconf
ffmpeg-python==0.2.0
spaces
moviepy
fastapi==0.111.0
requests
apscheduler==3.10.4
PyPubSub
pymysql==1.1.1
DBUtils==3.1.0
pika
retry
ffmpeg==0.4.0
